The first proposal of the circular linked list interface.
I tried to follow the naming principles from the project's wiki. This API is
highly influenced by the model of FreeBSD TAILQ queues. The only difference is
that this list is circular.

I'm not happy with this whole 'head type' approach. It makes it convoluted and
adds an additional argument to insertion methods. What I'd like to propose is
to ditch the head list concept at all. So we have only list elements, we do not
distinguish head. We do not provide the IS_HEAD, but IS_ELEM generalization. We
also get rid of APPEND, UNSHIFT and FIRST.

But this is ugly, because there is no natural representation of an empty list.
There can be so-called hacks, like making a "guard" element, or invalidating
next and prev pointers to NULL since in a 'normal' at least one-sized circular
list there is no situation where those pointers are NULL.

But I don't know whether we'd like to do that or not. Let me know what you think
about it. Maybe I'm missing something?