Add a scheduler.
I changed sched_yield to use a trap. So far this trap is handled
in an ugly way, but if this approach is accepted, I'll change it
to something nicer.
After an interrupt (or a trap), register $k1 stores the function that
should be executed: it can be either hardclock (in case of a regular
interrupt) or the sched_switch (in case of a trap caused by
sched_yield).

Function sched_switch() wasn't visible in intr.S, so I used a workaround
to make it link properly. I'll investigate the issue and fix it.

I also had some problems with making a label from intr.S visible in C
code.