Unconditionally notify waiting parent in proc_exit.
If the parent ignores SIGCHLD and waits for an only child process to exit,
the child won't notify the parent on exit, and the parent will
wait forever (unless interrupted somehow). POSIX says that in this
case, the parent's waitpid() call should fail with an error saying
that there were no children satisfying the criteria. If the child
notifies the parent unconditionally, then the parent will notice that
it has no children left matching the criteria, and the waitpid() call
will correctly fail.