Auto-reap fixes (#695)
* Unconditionally notify waiting parent in proc_exit.

If the parent ignores SIGCHLD and waits for an only child process to exit,
the child won't notify the parent on exit, and the parent will
wait forever (unless interrupted somehow). POSIX says that in this
case, the parent's waitpid() call should fail with an error saying
that there were no children satisfying the criteria. If the child
notifies the parent unconditionally, then the parent will notice that
it has no children left matching the criteria, and the waitpid() call
will correctly fail.

* Add check for presence of any children matching criteria in do_waitpid.

* Wait for the child in test_fork_sigchld_ignored.

wait() ensures that the child exits before the parent does,
which allows us to verify that SIGCHLD is indeed ignored.

* Notify parent's p_waitcv when reparenting children of exiting process.