Fix lost wakeup in do_waitpid().
It used to be that we always held all_proc_mtx while notifying the parent
of a child's state change. Now it's not always the case, so it's possible
to change a child's state and notify the parent before the parent waits on
its p_waitcv, resulting in a lost wakeup. We can fix it by introducing a flag
that indicates to the parent that some of its children have changed state.