Properly catch stop signals in sleepq_wait().
Sending a stop signal to a sleeping thread shouldn't interrupt
its sleep (in the sense of returning EINTR). It should make the
thread stop, and after it's resumed, return from sleepq_wait()
without an error, so it's somewhat like a spurious wakeup.
We expect callers to handle such wakeups correctly.