Allow fstat on file descriptor without read access.
Since we don't have read access to stdout, fstat() on STDOUT_FILENO would fail.
As a consequence, __swhatbuf() concludes that stdout isn't a tty, and we set
full buffering instead of line buffering on stdout.
The intended behavior is that stdout is line buffered by default if it's a tty.