Preserve value of x18 register across kernel interrupts.
Currently, the value of the x18 register can be changed if an interrupt occurs
in kernel mode. We fix this by not touching it in ctx_load.
Also, we simplify things by assuming in user_exc_leave that the user exception
frame is at the top of the stack. Previously this was not always the case
because kern_execve() directly called user_exc_leave(). Now, kern_execve() calls
switch_to_userspace(), which sets the stack pointer to the user exception frame.