build: fix KASAN build
sys/kern/Makefile do operations on KASAN variable before it is defined. As a
result following ifeq is always false:

> ifeq ($(KASAN), 1)
> SOURCES += kasan.c kasan_quar.c
> # KASAN's run-time library mustn't be sanitized!
> kasan.o : CFLAGS_KASAN =
> else
> # Clean dependencies without having to pass `KASAN=1` to `make clean`.
> CLEAN-FILES += .kasan.D .kasan_quar.D
> endif

The only solution for make KASAN build is passing KASAN=1 directly from cmdline.
We can avoid that by removing ifeq from Makefile and compiling kasan*.c always.

To do that we need to modify these .c files because they require to have
KASAN=1.

It's not the best solution but works with minimal effort.