cred: fix prototype of cred_can_access
This function is defined in sys/kern/cred_checks.c with the following signature:
> int cred_can_access(vattr_t *va, cred_t *cred, accmode_t mode)

The only place where it is called is vnode_access_generic where mode is
accmode_t.

This change requires small modyfication of includes because accmode_t is defined
in sys/vnode.h which also includes sys/cred.h.