Set window size on shell startup only if it isn't already set (#1130)
This prevent setwinsize from running when recording a terminal session with
script. We don't want it to run because when playing it back, the terminal will
send us an escape sequence containing the cursor position, but since setwinsize
isn't actually running (only the recording of its output), the escape sequence
gets echoed by the shell after finishing playback, which looks ugly.